Setup API route for the addition of Recipes
