--- server/mod/recipes.js
@@ -1,52 +1,0 @@
-const recipes = [
-  {
-    id: 1,
-    name: 'Banana milk shake',
-    description: 'Probably the best milk shake you ever had in your entire life',
-    category: 'Dessert',
-    ingredients: ['Milk', 'Banana', 'Olive oil'],
-    instructions: ['Blend the banana properly', 'Filter and shake'],
-    upVote: 0,
-    downVote: 0,
-    favorite: 0,
-    reviews: [
-      {
-        review: '',
-      },
-    ],
-  },
-  {
-    id: 2,
-    name: 'Banana milk shake',
-    description: 'Probably the best milk shake you ever had in your entire life',
-    category: 'Dessert',
-    ingredients: ['Milk', 'Banana', 'Olive oil'],
-    instructions: ['Blend the banana properly', 'Filter and shake'],
-    upVote: 0,
-    downVote: 0,
-    favorite: 0,
-    reviews: [
-      {
-        review: '',
-      },
-    ],
-  },
-  {
-    id: 3,
-    name: 'Banana milk shake',
-    description: 'Probably the best milk shake you ever had in your entire life',
-    category: 'Dessert',
-    ingredients: ['Milk', 'Banana', 'Olive oil'],
-    instructions: ['Blend the banana properly', 'Filter and shake'],
-    upVote: 0,
-    downVote: 0,
-    favorite: 0,
-    reviews: [
-      {
-        review: '',
-      },
-    ],
-  },
-];
-
-export default recipes;
