--- server/mod/review.js
@@ -1,12 +1,0 @@
-const review = [
-  {
-    id: 1,
-    content: 'An Amazing recipe. Really can\'t wait to try out this recipe with my family',
-  },
-  {
-    id: 2,
-    content: 'An Amazing recipe. Really can\'t wait to try out this recipe with my family',
-  },
-];
-
-export default review;
