--- server/mod/votes.js
@@ -1,12 +1,0 @@
-const votes = [
-  {
-    id: 1,
-    votes: 30,
-  },
-  {
-    id: 2,
-    votes: 100,
-  },
-];
-
-export default votes;
