+++ server/routes/index.js
@@ -1,6 +1,5 @@
-import Recipes from '../controllers/recipes';
 import userController from '../controllers/users';
-import recipeController from '../controllers/addRecipe';
+import recipeController from '../controllers/recipes';
 import Login from '../middleware/ensureLogin';
 import User from '../middleware/isuser';
 import recipeAdd from '../middleware/validateAddRecipe';
@@ -12,13 +11,4 @@ module.exports = (app) => {
   app.post('/api/v1/users/signup', userController.signUp);
   app.post('/api/v1/users/signin', userController.signIn);
   app.post('/api/v1/recipes/testAdd', Login.ensureLogin, User.isuser, recipeAdd, recipeController.addRecipe);
-
-
-  app.get('/api/v1/recipes', Recipes.getRecipes);
-  app.post('/api/v1/recipes', Recipes.createRecipe);
-  app.put('/api/v1/recipes/:recipeId', Recipes.updateRecipe);
-  app.delete('/api/v1/recipes/:recipeId', Recipes.deleteRecipe);
-  app.post('/api/v1/recipes/:recipeId/testVote', Recipes.upVote);
-  app.post('/api/v1/recipes/:recipeId/testDownVote', Recipes.downVote);
-  app.post('/api/v1/recipes/:recipeId/reviews', Recipes.reviews);
 };
